Changing the SetGradeAsync method and writing unit tests

+++ EP.Core/Entities/Grade.cs
@@ -6,6 +6,6 @@
         public string SubjectName { get; set; }
         public int StudentId { get; set; }
         public Student Student { get; set; }
-        public int Value { get; set; }
+        public byte Value { get; set; }
     }
 }
